+++ R/RcppExports.R
@@ -80,20 +80,21 @@ set_default <- function(input, value) {
 #' @param lead_time offset for expert forecasts. Defaults to 0, which means that
 #' experts forecast t+1 at t. Setting this to h means experts predictions refer
 #' to t+1+h at time t. The weight updates delay accordingly.
+#' @param allow_quantile_crossing Shall quantile crossing be allowed? Defaults to false which means that predictions are sorted in ascending order.
 #' @usage profoc(y, experts, tau, loss_function = "quantile",
 #' loss_parameter = 1, ex_post_smooth = FALSE, ex_post_fs = FALSE,
 #' lambda = -Inf, method = "boa", method_var = "A", forget_regret = 0,
 #' forget_performance = 0, fixed_share = 0, gamma = 1, ndiff = 1, deg = 3,
 #' knot_distance = 0.025, knot_distance_power = 1,
 #' gradient = TRUE, loss_array = NULL, regret_array = NULL,
-#' trace = TRUE, init_weights = NULL, lead_time = 0)
+#' trace = TRUE, init_weights = NULL, lead_time = 0, allow_quantile_crossing = FALSE)
 #' @return Profoc can tune various parameters automatically based on
 #' the past loss. For this, lambda, forget, fixed_share, gamma, ndiff,
 #' deg and knot_distance can be specified as numeric vectors containing
 #' parameters to consider. Profoc will automatically try all possible
 #' combinations of values provide.
 #' @export
-profoc <- function(y, experts, tau = as.numeric( c()), loss_function = "quantile", loss_parameter = 1, ex_post_smooth = FALSE, ex_post_fs = FALSE, lambda = as.numeric( c()), method = "boa", method_var = "A", forget_regret = as.numeric( c()), forget_performance = 0, fixed_share = as.numeric( c()), gamma = as.numeric( c()), ndiff = as.numeric( c()), deg = as.numeric( c()), knot_distance = as.numeric( c()), knot_distance_power = as.numeric( c()), gradient = TRUE, loss_array = as.numeric( c()), regret_array = as.numeric( c()), trace = TRUE, init_weights = NULL, lead_time = 0L) {
-    .Call(`_profoc_profoc`, y, experts, tau, loss_function, loss_parameter, ex_post_smooth, ex_post_fs, lambda, method, method_var, forget_regret, forget_performance, fixed_share, gamma, ndiff, deg, knot_distance, knot_distance_power, gradient, loss_array, regret_array, trace, init_weights, lead_time)
+profoc <- function(y, experts, tau = as.numeric( c()), loss_function = "quantile", loss_parameter = 1, ex_post_smooth = FALSE, ex_post_fs = FALSE, lambda = as.numeric( c()), method = "boa", method_var = "A", forget_regret = as.numeric( c()), forget_performance = 0, fixed_share = as.numeric( c()), gamma = as.numeric( c()), ndiff = as.numeric( c()), deg = as.numeric( c()), knot_distance = as.numeric( c()), knot_distance_power = as.numeric( c()), gradient = TRUE, loss_array = as.numeric( c()), regret_array = as.numeric( c()), trace = TRUE, init_weights = NULL, lead_time = 0L, allow_quantile_crossing = FALSE) {
+    .Call(`_profoc_profoc`, y, experts, tau, loss_function, loss_parameter, ex_post_smooth, ex_post_fs, lambda, method, method_var, forget_regret, forget_performance, fixed_share, gamma, ndiff, deg, knot_distance, knot_distance_power, gradient, loss_array, regret_array, trace, init_weights, lead_time, allow_quantile_crossing)
 }